Updated PickingSystem to use bitmasks
The picking system now uses bitmasks to determine what type of objects
(defined by UnitTypeComponent) can be hit when picking. Only the objects
eligible for picking will be considered. The bitmask is also prioritised
so that the item chosen when picking is the clostest of the hit objects
with the highest priority.