Adds a more rigorous method of setting component data.
I added a flag for to the graphics component called "render" to prevent
the flag from causing problems i made component data setting
a more rigourous deal, it now checks if you set all the values in one go
if you don't an exception with stack trace will be generated.

It's possible to skip the check by sending "true" as a fourth parameter
to entity set calls. The added error message causes a segfault sometimes
when triggered, the probable reason for this is the longjmp that
luaL_error does and how poorly this interacts with the C++
construct/destruct stack behaviour, this needs to be investigated
further.